requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-dotenv==1.0.0
pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.6
python-multipart==0.0.6

flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0

# Document processing libraries
pdfplumber==0.10.3
PyPDF2==3.0.1
PyMuPDF==1.23.8
pytesseract==0.3.10
opencv-python==4.8.1.78
ezdxf==1.1.4
pandas==2.1.4
numpy==1.25.2

# New dependencies for advanced processors
python-docx==1.1.0

requests==2.31.0

